[ home-UPDATE ] home페이지를 여러 컴포넌트로 분리하라 - 월 공연 리스트와 관련있는 date컴포넌트를 생성하였습니다 - codename에 따른 리스트를 보여주기 위해 codenameBtn컴포넌트를 생성하였습니다 - list,grid,calendar에 맞는 리스트 형식을 보여주기 위한 컴포넌트를 생성하였습니다 - 리스트 페이지에 api데이터만 들어있는 item컴포넌트를 생성하였습니다.
