+++ app/components/homeComp/HomeListItem.jsx
@@ -1,13 +1,11 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import styles from "../../styles/home.module.css";
-import { serviceKey } from "@/app/util/util";
-import { API_SortFunc, API_FilterFunc } from "./homeUtils";
+import { serviceKey,API_SortFunc, API_FilterFunc } from "@/app/util/utils";
 
 export default function HomeListItem({ codename }) {
   const [cultureList, setCultureList] = useState([]);
   const curDate = new Date().toISOString().split("T")[0];
-
   const url = `http://openapi.seoul.go.kr:8088/${serviceKey}/json/culturalEventInfo/1/30/
   ${codename ? codename : " "}/ /${codename === null ? curDate : " "}`;
