+++ app/components/layoutComp/layoutComp.module.css
@@ -66,7 +66,7 @@
 /* footer */
 .footer {
   width: 100%;
-  border-top: 1px solid teal;
+  /* border-top: 1px solid teal; */
   height: 150px;
   padding: 20px 0;
   box-sizing: border-box;
