--- app/util/util.js
@@ -1,1 +1,0 @@
-export const serviceKey = process.env.NEXT_PUBLIC_SERVICEKEY;
